+++ src/calibre/utils/lzx/lzxmodule.c
@@ -152,7 +152,7 @@ decompress(PyObject *self, PyObject *args)
 {
     unsigned char *inbuf;
     unsigned char *outbuf;
-    unsigned int inlen;
+    Py_ssize_t inlen;
     unsigned int outlen;
     int err;
     memory_file source;
